fixed settings for threeD bar depth - demonstrating in Bars example
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/trunk/kdchart@361 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ src/KDChartAbstractThreeDAttributes.cpp
@@ -102,13 +102,13 @@ int AbstractThreeDAttributes::height() const
     return d->height;
 }
 
-void AbstractThreeDAttributes::setDepth( int depth )
+void AbstractThreeDAttributes::setDepth( double depth )
 {
     d->depth = depth;
 }
 
 
-int AbstractThreeDAttributes::depth() const
+double AbstractThreeDAttributes::depth() const
 {
     return d->depth;
 }
